+++ tests/ModelTest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Lioneagle\LioneaglePaginator\Tests;
+
+/**
+ * @internal
+ * @coversNothing
+ */
+class ModelTest extends TestCase
+{
+    public function testThePaginatorReturnsTheCorrectNumberOfRecords()
+    {
+        $pageSize = config('lioneagle-paginator.page_size.default');
+
+        $query = Model::paginator();
+
+        $this->assertCount($pageSize, $query);
+    }
+}